Move API endpoint under `/storage`
