+++ backend/infrahub/api/__init__.py
@@ -14,7 +14,6 @@ from infrahub.api import (
     auth,
     diff,
     file,
-    file_object,
     internal,
     menu,
     oauth2,
@@ -36,7 +35,6 @@ router.include_router(artifact.router)
 router.include_router(auth.router)
 router.include_router(diff.router)
 router.include_router(file.router)
-router.include_router(file_object.router)
 router.include_router(internal.router)
 router.include_router(menu.router)
 router.include_router(oauth2.router)
